implement read_text_to_string
